HamSCI_NewChirp/requirements.txt
numpy
scipy.signal
matplotlib.pyplot
scipy.io.wavfile
